pyproject.toml
[project]
name = "python-full-stack"
version = "0.1.0"
description = ""
authors = [
    {name = "pankesh.patel",email = "pankesh.patel@bolt-data.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (>=1.7.4,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"

# command to run - poetry add <package-name> --source gitlab 
httpx = "^0.28.1"
[project.optional-dependencies]
all = []



# command to run poetry add <package-name> --source gitlab 
# [[tool.poetry.source]]
# name = "<source-name>"
# url = "<source-url>"
# priority = "default"  # or "primary", "supplemental", "explicit"

# example
# [[tool.poetry.source]]
# name = "gitlab"
# url = "https://gitlab.example.com/api/v4/projects/123/packages/pypi/simple"
# priority = "primary"


# [tool.pyright]
# [tool.ruff.lint.isort.sections]
# [tool.ruff.lint.isort]
# [tool.ruff.lint.pycodestyle] 
# [tool.ruff]
# [tool.mypy]
#[tool.pytest.ini_options]
#[tool.black]